+++ src/main/java/com/example/programame_project_api/entities/SimpleDonation.java
@@ -12,7 +12,7 @@ public class SimpleDonation {
     private  long id;
 
     @Column(nullable = false)
-    private float amount;
+    private double amount;
 
 
     @OneToOne(  mappedBy = "simpleDonation", fetch = FetchType.LAZY)
@@ -21,7 +21,7 @@ public class SimpleDonation {
     public SimpleDonation() {
     }
 
-    public SimpleDonation(float amount) {
+    public SimpleDonation(double amount) {
         this.amount = amount;
 
     }
@@ -30,11 +30,11 @@ public class SimpleDonation {
         return id;
     }
 
-    public float getAmount() {
+    public double getAmount() {
         return amount;
     }
 
-    public void setAmount(float amount) {
+    public void setAmount(double amount) {
         this.amount = amount;
     }
